% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_js_files_to_tempfile.R
\name{add_code_menu_to_js_tempfile}
\alias{add_code_menu_to_js_tempfile}
\title{add_code_menu_to_js_tempfile}
\usage{
add_code_menu_to_js_tempfile(r_options)
}
\arguments{
\item{code_folding}{code_folding = { default = 'show' | NULL, 'hide', 'off' }}

\item{code_download}{code_download = { default = TRUE | NULL, FALSE | 'off' }}
}
\value{
NULL or js_tempfile
}
\description{
add_code_menu_to_js_tempfile
}
