% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_parg_from_res}
\alias{build_parg_from_res}
\title{Build Pandoc argument from resources}
\usage{
build_parg_from_res(res_name, res_folder, res_file)
}
\value{
String type.
}
\description{
Build Pandoc argument from resources
}
