% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_embed.R
\name{embed_file_a}
\alias{embed_file_a}
\title{Embed a file to HTML throght a tag}
\usage{
embed_file_a(file, name = basename(file), text = paste("Download", name), ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Embed a file to HTML throght a tag
}
