% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pathfile_from_res}
\alias{get_pathfile_from_res}
\title{Get path and file name from resources}
\usage{
get_pathfile_from_res(res_folder, res_file)
}
\value{
String type.
}
\description{
Get path and file name from resources
}
