% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_merge}
\alias{list_merge}
\title{Merge 2 list}
\usage{
list_merge(a, b)
}
\arguments{
\item{a}{List type, can \code{unnamed}.}

\item{b}{List type, will ignore \code{unnamed}.}
}
\value{
Output list type was merged.
}
\description{
Merge 2 list into 1 list.
}
\examples{
l1 <- list(a = T, b = F, d = list(g = 3))
l2 <- list(b = T, c = T, d = list(e = 1, f = 2))
list_merge(l1, l2)
}
\seealso{
\itemize{
  \item \href{https://github.com/tidyverse/purrr/blob/master/R/list-modify.R}{https://github.com/tidyverse/purrr/blob/master/R/list-modify.R}
}
}
