% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_list}
\alias{print_list}
\title{Print tree member of list type}
\usage{
print_list(indent_string = "", lst = NULL)
}
\arguments{
\item{indent_string}{String type.}

\item{lst}{List type.}
}
\value{
Nothing.
}
\description{
Print tree member of list type
}
