% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rb3m-package.R
\docType{package}
\encoding{UTF-8}
\name{Rb3m-package}
\alias{Rb3m-package}
\title{Rb3m package}
\source{
\itemize{
    \item {\href{https://github.com/BabyMouse/Rb3m}{https://github.com/BabyMouse/Rb3m}}
  }
}
\description{
Rb3m is a template for RMarkdown.
}
\seealso{
\itemize{
    \item {\href{https://github.com/BabyMouse/Rb3m/blob/master/README.md}{https://github.com/BabyMouse/Rb3m/blob/master/README.md}}
  }
}
\author{
\describe{
    \item{}{Trát Quang Thụy}
  }
}
