% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Robot-helper-template.R
\name{RH-template-package}
\alias{RH-template-package}
\title{Robot helper template}
\source{
\itemize{
    \item \href{https://cran.r-project.org/web/packages/roxygen2/vignettes/rd.html}{https://cran.r-project.org/web/packages/roxygen2/vignettes/rd.html}
  }
}
\usage{
add_numbers(x)
add_numbers(x, y)
}
\arguments{
\item{x}{A description of the parameter 'x'. The
description can span multiple lines.}

\item{y}{A description of the parameter 'y'.}
}
\value{
Return or output of function \code{add_numbers}.
}
\description{
A robot helper template for roxygen2.

Add some features

A robot helper template for roxygen2.
}
\details{
% Comment some things...


Details for some things.
}
\note{
\enumerate{
    \item Note 1.
    \item Note 2.
  }
}
\section{A Custom Section}{

Text accompanying the custom section.
}

\examples{
add_numbers(1, 2) ## returns 3

## don't run this in calls to 'example(add_numbers)'
\dontrun{
   add_numbers(2, 3)
}

## don't test this during 'R CMD check'
\donttest{
   add_numbers(4, 5)
}

}
\references{
\describe{
    \item{label-1}{text-1}
    \item{label-2}{text-2}
  }
}
\seealso{
\itemize{
  \item \href{https://github.com/BabyMouse/Rb3m}{GitHub: https://github.com/BabyMouse/Rb3m}
  \item \href{http://r-pkgs.had.co.nz/man.html}{http://r-pkgs.had.co.nz/man.html}
}
}
\author{
\describe{
    \item{}{Trát Quang Thụy}
  }
}
