% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_list}
\alias{print_list}
\title{Print tree member of list type}
\usage{
print_list(lst = NULL, indent_string = NULL, list_title = NULL)
}
\arguments{
\item{lst}{List type. Default = NULL.}

\item{indent_string}{String type. Default = "".}

\item{list_title}{String type. Default = "".}
}
\value{
Nothing.
}
\description{
Print tree member of list type
}
